// src/assets/language.d.ts
export interface LanguageData {
  kor: {
    about: string;
    start: string;
    newInvest: string;
    newglobal: string;
    adventure: string;
    create: string;
    trust: string;
    value: string;
  };
  eng: {
    about: string;
    start: string;
    newInvest: string;
    newglobal: string;
    adventure: string;
    create: string;
    trust: string;
    value: string;
  };
}